Update corewatcher to use autotools.
This involves major source restructuring and in the process
some files have been renamed as well as moved.  This was
done to ease the use of autotooling the package though
it shouldn't impair clarity of purpose for the files.

At this point there are further opportunities to improve
the packaging such as introducing #defines for some
file locations in the code.

Signed-off-by: William Douglas <william.douglas@intel.com>