Add file flag to corewatcher.
Allow corewatcher to be given an optional file argument.  This file
will be a trace file generated by corecatcher in the /tmp folder.
Corewatcher will then use this file to submit a crash and then exit
instead of polling and waiting for a crash to submit.

Usage: corewatcher -f filename
Signed-off-by: William Douglas <william.douglas@intel.com>