Add script for submitting crash file.
At this point we will no longer be using the submit function built
into corewatcher.  Instead a submit will built into a shell script
instead and used in place of corewatcher's submit.  This script is
currenctly a stub however and will need real code put in place for
running the submit.

Signed-off-by: William Douglas <william.douglas@intel.com>