Rework how we expect to scan the core dump folder.
scan_corefolders now will take the trace file as an argument
regardless of how corewatcher was called.  This is done as the
preliminary step before inserting the epoll logic.  Also since we are
getting trace file names instead of core files, we should fix up the
get_trace_file function into a get_core_file one.  Also since we may
have a trace file before we get a core file and we shouldn't need to
submit the core file.  Try and process through the trace file first
and once we submit backup the core file if possible.

Signed-off-by: William Douglas <william.douglas@intel.com>