Do polling for crashes with epoll.
This patch introduces polling a file descriptor that will be updated
with the name of the trace file by corecatcher. This implementation
is intended to make it fairly simple to add in systemd support.

We have to move a few things around and I took the chance to try and
add some labels for failure during main. There is quite a bit in main
that should still be relocated however.

Signed-off-by: William Douglas <william.douglas@intel.com>