Remove extra unlink
This was the unlink of the actually processed file.  The primary duplicate
detection is in move_core().  Due to the two phases of processing in
scan_corefolder() it is a fully normal path for a core to get added
to the work queue and then a second add attempt likely happens at
reprocess_corefile().

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>