Add inotify support, remove pthreads dep, bump version
Arjan noted a lot of activity on wake/poll.  This was when the server
was down and cores/reports were present to submit so there was work
to do.  But a lot of effort was generated to no gain.  The client should
be smarter with respect to sending reports on demand and minimizing
resource usage.  We still need additional state, parallelization and event
connectors to allow different threads to prepare reports for submission,
improve current report quality if high speed network (and thus debuginfo)
becomes available, do just report submission if low quality network
becomes available, etc.

This commit adds inotify support so we immediately attempt to process
cores and reduces the poll loop time to 15 minutes.  Lacking more
sophisticated network event driven logic, a slow poll loop can still
lead to reports eventually being submitted if they weren't able to be
fully submitted at inotify time.

In support of this a few functions are split out and GThreads/GMutexes
used instead of pthread.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>