Add timestamp from core
The server has been attributing a timestamp to a crash report based on the
time the report is received.  This may differ greatly from the actual time
of the issue.  This is easily fixed with a stat of the corefile.

Unfortunately, the same can't as easily be done for getting the os-build,
which may also be incorrectly attributed on the client side based on the
/etc/os-release file's contents at the time of a crash's processing by
corewatcher, as opposed to at the actual time of the crash.

Tracking the combination of the reported os_build, core time, and receiving
time should give us more rich reporting and ability to discern problems
between different classes of users (eg: more/less network connectedness,
more/less aggressive adoption of updates).

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>