Always chmod our directories.
I observed the the modes of corewatcher's directories had changed on one
of my systems.  I don't know what would have caused that, but it led to
cores not being recorded in the filesystem.  To be robust we create the
directories if for some reason they do not exist at start up.  Similarly,
try to always insure they have the correct permissions at start up.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>