Fix core_pattern being unset
In the commited version I lost a multiply by 100 that'd previously had.
I'd intended for the free metric to be a percentage represented by an
integer with value 0..100.  This error means the core_pattern was unset
just at the first call to scan_folder().  That is corrected here.

Additionally the core_pattern is explicitly set on start now.

And the log messages are a bit more clear.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>