Hard code system path
There is little point reading the path from the environment.

First, it will be the path which systemd set for the environment in which
systemd started the corewatcher daemon.  That may or may not match the
system path.  And it is even more likely to not match the path set in
the environment in which a regular user has interactively started an app.

Second, by searching the path, we get false positive matches.  The crashdb
server has shown a variety of examples of crashes where a matching
app name was found in a developer's custom path.  We also have gotten
crashes reports where a developer was working on local dev/test of a
program whose name matched a binary's name from elsewhere in the path
(subsequent patches will deal more with this case).

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>