Clean path searching for binary
Gdb needs an executable and a core so we seek out an executable which
may correspond to the crash.  This was done in a loose way and led to
false matches.  This patch makes the search more strict, adds some debug
message output and handles errors better.  As per the earlier commit,
the "path" is hard coded and with only a single member ("/usr/bin"),
but the code remains to allow it to be a colon separated list of
directories.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>