Correct file renaming on skip
With previous commits' refactoring the process_common() function is really
part of just create_report().  Pulling it into one function makes it easier
to handle the skip cases which need to do different renames depending on
whether we're processing or re-processing.  I'd previously used move_core()
for skips, but that really is tailored to the first state transition from
core dumped to core captured by corewatcher to process.  So add a
skip_core() function and use it.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>