Search precisely for gdb warnings
There are many reasons gdb might print a "warning: " string.  There are two
that I know we care about.  It might additionally make sense to watch for
messages like:

warning: .dynamic section for "/usr/lib64/libsomething.so.42" is not at the
expected address (wrong library or version mismatch?)

warning: Could not load shared library symbols for 2 libraries, e.g.
libsomething.so.42

For those libraries it might make sense to annotate the backtrace so
the crashdb could decrement its quality.  Or in the face of these it
might make sense to simply declare the backtrace as bad.  I'm inclined
to think these traces should be sent, but with as much information as
possible about the apparent quality issues.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>